requirements.txt
aiogram>=3.15
sqlalchemy[asyncio]>=2.0.36
asyncpg
aiosqlite
python-dotenv
google-generativeai>=0.8.0
